requirements.txt
flask
flask-cors
openai
python-dotenv
numpy
scikit-learn
gunicorn
